refactor: useLibrarySettings hook remove unused useCategories hook
