+++ src/hooks/index.ts
@@ -2,7 +2,6 @@ export { default as useSearch } from './common/useSearch';
 export { default as useFullscreenMode } from './common/useFullscreenMode';
 export { default as useNovelTrackerInfo } from './useNovelTrackerInfo';
 export { default as useBoolean } from './common/useBoolean';
-export { useCategorySettings, useLibrarySettings } from './useSettings';
 export { usePreviousRouteName } from './common/usePreviousRouteName';
 export { useBackHandler } from './common/useBackHandler';
 export { useTextToSpeech, TextToSpeechStatus } from './common/useTextToSpeech';
