+++ src/hooks/persisted/index.ts
@@ -2,3 +2,4 @@ export { useTheme } from './useTheme';
 export { useUpdates, useLastUpdate } from './useUpdates';
 export { default as useCategories } from './useCategories';
 export { default as useHistory } from './useHistory';
+export { useLibrarySettings } from './useSettings';
