+++ src/screens/library/components/LibraryBottomSheet/LibraryBottomSheet.tsx
@@ -3,8 +3,7 @@ import { StyleSheet, Text, useWindowDimensions, View } from 'react-native';
 import { SceneMap, TabBar, TabView } from 'react-native-tab-view';
 import color from 'color';
 
-import { useLibrarySettings } from '@hooks/useSettings';
-import { useTheme } from '@hooks/persisted';
+import { useLibrarySettings, useTheme } from '@hooks/persisted';
 import { getString } from '@strings/translations';
 import { Checkbox, SortItem } from '@components/Checkbox/Checkbox';
 import {
