+++ src/screens/more/MoreScreen.tsx
@@ -7,13 +7,12 @@ import {
 } from 'react-native';
 import { Switch } from 'react-native-paper';
 import { useSelector } from 'react-redux';
-import { getString } from '../../../strings/translations';
+import { getString } from '@strings/translations';
 
 import { List } from '@components';
 
 import { MoreHeader } from './components/MoreHeader';
-import { useLibrarySettings } from '@hooks/useSettings';
-import { useTheme } from '@hooks/persisted';
+import { useLibrarySettings, useTheme } from '@hooks/persisted';
 import { MoreStackScreenProps } from '@navigators/types';
 import { RootState } from '@redux/store';
