+++ src/screens/settings/SettingsGeneralScreen/modals/DisplayModeModal.tsx
@@ -7,9 +7,9 @@ import { Text, StyleSheet } from 'react-native';
 
 import { Portal, Modal, overlay } from 'react-native-paper';
 
-import { RadioButton } from '../../../../components/RadioButton/RadioButton';
-import { useLibrarySettings } from '@hooks/useSettings';
+import { RadioButton } from '@components/RadioButton/RadioButton';
 import { ThemeColors } from '@theme/types';
+import { useLibrarySettings } from '@hooks/persisted';
 
 interface DisplayModeModalProps {
   displayMode: DisplayModes;
