+++ src/screens/settings/SettingsGeneralScreen/modals/GridSizeModal.tsx
@@ -3,10 +3,10 @@ import { Text, StyleSheet } from 'react-native';
 
 import { Portal, Modal, overlay } from 'react-native-paper';
 
-import { RadioButton } from '../../../../components/RadioButton/RadioButton';
+import { RadioButton } from '@components/RadioButton/RadioButton';
 
-import { useLibrarySettings } from '@hooks/useSettings';
 import { ThemeColors } from '@theme/types';
+import { useLibrarySettings } from '@hooks/persisted';
 
 interface GridSizeModalProps {
   novelsPerRow: number;
